executor: refactor embedded syslog launching code
Not sure why LaunchSyslogServer hasn't always been in executor.go. Also
cleaned up a comment.
